requirements.txt
glfw==1.6.0
numpy==1.14.3
PyOpenGL==3.1.0
typing==3.6.4
healpy==1.13.0
trimesh==3.5.6
freetype-py==2.1.0.post1
